Switching to the new toolset compiler.
The new version is
https://dotnet.myget.org/feed/roslyn/package/nuget/Microsoft.Net.Compilers/2.3.0-rdonly-ref-61607-04

The new version enforces the following rules:
- `Span<T>` and `ReadonlySpan<T>` are stack-only types. (no boxing, invalid use in async, etc...)
- `Span<T>` and `ReadonlySpan<T>` cannot be passed via a writeable parameter.

There were, however, no violations of the rules. (I had to introduce and then remove some violatins, just to se that detection works)

NOTE: we are not yet validating the usage of span-containing structs. There could be more indrect violations of stack safety through that.