Fix outgoing back pressure stalling closing the socket
- This change does a few things:

1. It adds the events we will replace with
pipe events to IConnectionContext and IConnectionInformation to get out of
band notifications about pipe completions.

2. It also implements those callbacks
and exposing slight changes we'll need to make once we have them. The idea is
that we can delete/replace these methdods once we have the new pipe API and things
will keep working

3. It fixes an issue where sockets potentially wouldn't close because of
output back pressure from the OS. If there was no pending pipe.ReadAsync call the
signal to cancel the pending read could be missed.