Fix for #2187 (#2180)
- Changed the ReadableBufferReader's Cursor computation to avoid special
casing end.
- The issue was caused by a mix of things and the interaction with pipes.
The ReadableBufferReader was moving to the next block but also skipping
the _index update since it special cases when you've reached the end of the
buffer. As a result, the cursor ended up pointing to the end of the empty
block and the pipe ended up having a negative length for the buffer.